Fixed minor issues and mistakes

+++ src/Services/EventActionService.cpp
@@ -6,7 +6,7 @@ void EventActionService::addEventActionDefinitions(Message message){
 	// TC[19,1]
 	if (message.messageType == 1 && message.packetType == Message::TC && message.serviceType
 	                                                                     == 19){
-		int N = message.readUint16();
+		uint16_t N = message.readUint16();
 		for (uint16_t i = 0; i < N; i++){
 
 		}
@@ -18,7 +18,7 @@ void EventActionService::deleteEventActionDefinitions(Message message) {
 	// TC[19,2]
 	if (message.messageType == 2 && message.packetType == Message::TC && message.serviceType
 	                                                                     == 19){
-		int N = message.readUint16();
+		uint16_t N = message.readUint16();
 		for (uint16_t i = 0; i < N; i++){
 
 		}
@@ -38,7 +38,7 @@ void EventActionService::enableEventActionDefinitions(Message message) {
 	// TC[19,4]
 	if (message.messageType == 4 && message.packetType == Message::TC && message.serviceType
 	                                                                     == 19){
-		int N = message.readUint16();
+		uint16_t N = message.readUint16();
 		for (uint16_t i = 0; i < N; i++){
 
 		}
@@ -50,7 +50,7 @@ void EventActionService::disableEventActionDefinitions(Message message) {
 	// TC[19,5]
 	if (message.messageType == 5 && message.packetType == Message::TC && message.serviceType
 	                                                                     == 19){
-		int N = message.readUint16();
+		uint16_t N = message.readUint16();
 		for (uint16_t i = 0; i < N; i++){
 
 		}
